Random Code Go
Miscellaneous Robotics code that accomplishes various tasks throughout the engineering process of building and programming the robots. Among the important code is the CameraDetection.c and Camera_to_Pick_Up_Ball.c. The former detects an object in a plane using the camera. The latter uses the camera to create a sweeping robot that checks for the object, moves towards it, centers the claw with the object, grabs the object using a distance sensor, picks up the object with claws, and throws object when Analog push sensor is pushed.